add partial handling of slide end time
