+++ src/providers/common/index.ts
@@ -1,0 +1,3 @@
+export {HotspotLoader} from './hotspot-loader';
+export {ThumbLoader} from './thumb-loader';
+export {ThumbUrlLoader} from './thumb-url-loader';
